use .mjs extension
